+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [0.74.0]
 
 * Add refactoring support for converting a Prolog module to an object
+* Fix predicate refactorings to not be offered for conditional compilation directives
 
 ## [0.73.0]